+++ source/ChanSort.Loader.Samsung/DigitalChannel.cs
@@ -21,9 +21,9 @@ namespace ChanSort.Loader.Samsung
       int transp = data.GetByte(_ChannelOrTransponder);
       decimal freq = transpFreq.TryGet(transp);
       if (freq == 0)
-        freq = transp*8 + 106; // (106 = DVB-C; DVB-T=306?)
+        freq = LookupData.Instance.GetDvbtFrequeny(transp); // transp*8 + 106); // (106 = DVB-C; DVB-T=306?)
 
-      this.ChannelOrTransponder = LookupData.Instance.GetDvbtTransponder(freq).ToString();
+      this.ChannelOrTransponder = transp.ToString();
       this.FreqInMhz = freq;
     }
   }