+++ source/changelog.md
@@ -1,6 +1,11 @@
 ChanSort Change Log 
 ===================
 
+2015-05-06
+- added support for Samsung J-Series DVB-T and analog channel lists
+- fixed reference lists with Samsung J-Series
+- web site and source code moved to http://github.com/PredatH0r/ChanSort
+
 2015-04-18
 - added support for Samsung J-Series DVB-C and DVB-S channel lists
   (analog channels and DVB-T are not supported yet)